pyproject.toml
#################################################
# Project metadata
#################################################

[project]
requires-python = ">=3.11"
name = "temple8"
version = "0.1.0"
description = "A minimal FastAPI project template."
readme = "README.md"
keywords = ["fastapi", "template", "minimal", "docker", "alembic", "sqlalchemy", "celery", "rabbitmq", "grpc", "circuit-breaker", "postgres"]
license = {file = "LICENSE"}
authors = []
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.0",
    "black>=24.10.0",
    "celery[postgresql]>=5.4.0",
    "circuitbreaker>=2.0.0",
    "fastapi>=0.115.4",
    "greenlet>=3.1.1",
    "grpc-interceptor>=0.15.4",
    "grpcio-tools>=1.67.1",
    "grpcio>=1.67.1",
    "gunicorn>=23.0.0",
    "isort>=5.13.2",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-json-logger>=2.0.7",
    "python-multipart>=0.0.17",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn>=0.32.0",
]
[dependency-groups]
dev = [
    "httpx>=0.27.2",
    "mypy>=1.13.0",
    "pip-tools>=7.4.1",
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "requests>=2.32.3",
    "ruff>=0.7.1",
]

#################################################
# Mypy config
#################################################

[tool.mypy]
follow_imports = "skip"
ignore_missing_imports = true
warn_no_return = true
warn_unused_ignores = true
allow_untyped_globals = true
allow_redefinition = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "svc.tests.*"
ignore_errors = true

#################################################
# Ruff config
#################################################

[tool.ruff]
respect-gitignore = true
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes `E` and `F` codes by default
select = ["E", "F", "PT", "C4", "I"]
ignore = ["E501"]

per-file-ignores = {}

[tool.ruff.lint.mccabe]
max-complexity = 10

#################################################
# Pytest config
#################################################

[tool.pytest.ini_options]
addopts = "--strict-markers --maxfail 1 --cov svc tests/ --no-header"
markers = """
    integration: mark a test as an integration test.
"""
console_output_style = "progress"

#################################################
# Setuptools config
#################################################

[tool.setuptools.packages.find]
where = ["app"]  # ["."] by default
